+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 This new release includes :
+- serialization support,
+- calendar API improvements (composite calendars),
+
+This new release also includes breaking changes :
+- deprecated methods removal (see details below).
 
 ### Added
 
@@ -19,5 +24,11 @@ This new release includes :
 ### Fixed
 
 ### Deprecated
+
+### Removed
+
+- Remove deprecated method `IsoCountry.getCode()` : use `IsoCountry.getAlpha2Code()` instead (#49).
+- Remove deprecated method `IsoCountry.fromCode(String)` : use `IsoCountry.fromAlpha2Code(String)`
+  instead (#49).
 
 ### Thanks